Fix issue with operator always pulling the latest
We are defaulting this variable to a blank string, which means the default value specify is skipped, as its only for undefined variables.  This caused it to use just /awx-operator for the image name.  Since the image path is now invalid, it just ignored the image version we specified and pulled the latest operator version.